Adding author to OriginalText and formatting

+++ OldLanguages/templates/originaltexts.html.php
@@ -1,6 +1,6 @@
-<p><?=$totalOriginalTexts?> original texts have been submitted to the Internet Old Text Database.</p>
+<p><?= $totalOriginalTexts ?> original texts have been submitted to the Internet Old Text Database.</p>
 <table>
-<tr>
+  <tr>
     <th>Title</th>
     <th>Original Text</th>
     <th>Orig. Text Photo</th>
@@ -9,34 +9,36 @@
     <th>Date</th>
     <th>Actions</th>
   </tr>
-<?php foreach ($originalTexts as $originalText): ?>
-  <tr>
-  <?php 
-  $date = date_create($originalText->insert_date);
-  echo "<td hidden>".htmlspecialchars($originalText->id, ENT_QUOTES, 'UTF-8')."</td>"
-  ."<td>".htmlspecialchars($originalText->title, ENT_QUOTES, 'UTF-8')."</td>"
-  ."<td>".htmlspecialchars($originalText->text, ENT_QUOTES, 'UTF-8')."</td>"
-  ."<td>"
-  .'<img class="autoResizeImage" src="data:image/jpeg;base64,'.base64_encode($originalText->text_img).'"/>'
-  ."</td>"
-  ."<td>".htmlspecialchars($originalText->getPlace()->place, ENT_QUOTES, 'UTF-8')."</td>"
-  ."<td>".htmlspecialchars($originalText->getOldLanguage()->language, ENT_QUOTES, 'UTF-8')."</td>" 
-  ."<td>".htmlspecialchars(date_format($date,"d-m-Y"), ENT_QUOTES, 'UTF-8')."</td>";
-  ?>
-  <td>
-  <a class="details" href="/originaltext/edit/<?=$originalText->id?>">Edit</a>
-  <br/>
-  <a class="details" href="/originaltext/delete/<?=$originalText->id?>">Delete</a>
-  <td>  
-  </tr>
-<?php endforeach; ?>
+  <?php foreach ($originalTexts as $originalText): ?>
+    <tr>
+      <?php
+      $date = date_create($originalText->insert_date);
+      echo "<td hidden>" . htmlspecialchars($originalText->id, ENT_QUOTES, 'UTF-8') . "</td>"
+        . "<td>" . htmlspecialchars($originalText->title, ENT_QUOTES, 'UTF-8') . "</td>"
+        . "<td>" . htmlspecialchars($originalText->text, ENT_QUOTES, 'UTF-8') . "</td>"
+        . "<td>"
+        . '<img class="autoResizeImage" src="data:image/jpeg;base64,' . base64_encode($originalText->text_img) . '"/>'
+        . "</td>"
+        . "<td>" . htmlspecialchars($originalText->getPlace()->place, ENT_QUOTES, 'UTF-8') . "</td>"
+        . "<td>" . htmlspecialchars($originalText->getOldLanguage()->language, ENT_QUOTES, 'UTF-8') . "</td>"
+        . "<td>" . htmlspecialchars(date_format($date, "d-m-Y"), ENT_QUOTES, 'UTF-8') . "</td>";
+      ?>
+      <?php if (!(empty($user) || $user->id == $originalText->getAuthor()->id)): ?>
+        <td>
+          <a class="details" href="/originaltext/edit/<?= $originalText->id ?>">Edit</a>
+          <br />
+          <a class="details" href="/originaltext/delete/<?= $originalText->id ?>">Delete</a>
+        <td>
+        <?php endif; ?>
+    </tr>
+  <?php endforeach; ?>
 </table>
 Select page:
 
 <?php
 for ($i = 1; $i <= $numPages; $i++):
 ?>
-  <a href="/originaltext/list/<?=$i?>"><?=$i?></a>
+  <a href="/originaltext/list/<?= $i ?>"><?= $i ?></a>
 <?php endfor; ?>
 
 </div>
